# [Xamarin.Android.Build.Tasks] perf improvements for <ResolveAssemblies/>